added separate class for signal model
