+++ DAL/Intersection_Model.py
@@ -1,0 +1,6 @@
+class IntersectionModel:
+    def __init__(self, _id: str, signals: list):
+        self._id = _id
+        self.signals = signals
+
+
