+++ DAL/Signal_Model.py
@@ -1,9 +1,3 @@
-class IntersectionModel:
-    def __init__(self, _id: str, signals: list):
-        self._id = _id
-        self.signals = signals
-
-
 class SignalModel:
     def __init__(self, _id: str, link: str):
         self._id = _id
